The "initialising system" window, can we change it back to "TOD" instead of "system"? And on a wider viewport like my laptop it is centered correctly. with a narrow viewport to simulate a mobile display, "Initialising System" is still aligned left. And can we extend the connection window to 2000 ms please and replace the lightning icon with a wifi or signal icon, and reduce the icon size somewhat.
On the team selection window, why are we not changing the global theme when a tile is selected? I thought that would simplify things a lot?

Currently, after the last changes, the only indication that a tile is selected is that a button appears, but there is no extra border appearing and the backgrounds don't show any change. The borders and the button remain green. Clicking on confirm, the codename window opens, but it's title is white and the glow effects are white, not the team blue. Closing the codename window, the Select your allegiance title is also now white and the glow effects are also white, and hovering over the observer option also white. Unselecting all tiles does not return the color to green.

Once I'm on the tod, all the glows are white also.

Won't it be considerably easier just to change the global theming or at least the green/blue/red value there and then as a tile is selected or deselected? So when selected is clicked, the codename window can just be displayed?

The buttons only appear when a tile is selected, so will it be useful to have the buttons permanently set to the team's colors, instead of passing explicitTheme?