Remove the dark background from the first div, and give the PAD div the dark background. Right now the PAD title area is in the middle of the screen as opposed to just be above the fold, as if the title area's autofill height didnt take effect. So the pad title should be just above the fold. The top div will autofill the remaining height and cover the majority of the page, right now the content of the top div just stacks, bringing the content below it (the pad) higher up the window. Within the top div, the stats area will stack normally and the title area will auto fill the remaining height. If it helps to create two divs within the first div to apply height attributes individually that's fine, visually it will appear the same. And just to be clear, the top div is at the bottom, and as the PAD div slides up on top of the first div, the pad remains visible as the first div is hidden below it.
The pad's behaviour seems to work right, it's just that the PAD switches on and displays the info area when it reaches the bottom of the first div, ie. a user barely starts scrolling, whereas id like it to switch on when reaching the top of the screen. Is it possible to apply a magnet effect to the scrolling pad, that the Pad's title bar/button bar magnets to either the top or the bottom position, and is it worth considering implementing? The PAD is of no use in the halfway position, and it will be blank for most of the page as the info area will only be displayed when the pad is in the top position, so a slight slide or nudge will magnet the pad to the opposite position?