No, please undo the last edits. Our layout changed and it's not perfect anymore.
Then in the past I asked you to have the button bar only scroll over the stats area, but since we've made the first div take up the majority of the screen I think I requested that the pad can scroll to the top of the screen, as opposed to below the agent name and team name, which should make things less complicated.

So can we undo the last edit first? It's important to me that we not loose that layout. Then I'll prompt you further from there.