Our pad has blue borders, rounded corners, but the background is not transparent. I think we've narrowed it down to it being a browser issue. Thank you for your patience and trying every angle. I find it odd, but I'm not a programmer, that a basic function like transparency gives such problems. I mean, how do actual programmers manage to build something when there are these quirks popping up? If I was a programmer, it would have taken me days/weeks to try all the options we've tried, and that's assuming I know all the various ways to set transparency, and then *not* find a solution, that would be very very frustrating. I'm glad I have you to help me.
Okay, have we've narrowed it down to the transparency that breaks the background, and if we've tried all the options then this is the best  we can do. I accept that there are problems that even the great Gemini that knows just about everything knowable can't resolve, unless you change the way browsers behave  to solve the problem on that side, but that's out of our control.
So with a non-transparent background, the backdrop blur would be of no use so we can remove that, and we can just apply the gloss. But before we do, I'd like to apply the gloss to the pad's info area only, so it appears like the pad's screen has a reflection. Is that possible?

I'm just curious, with my tasker understanding: Some elements transparency does work. In Tasker, so I would try to duplicate the element that does work to solve the problem, so in this case, I would copy the info area, change its colors and values and set it on top of the pad's div, behind the info area, so essentially making the original div just an empty container/holder for the new copied and styled area. But in this case, i assume that slapping on a plaster like this might not be the most ideal or optimum way to do things.

But if we've tried everything, then let's move on with styling another section of the TOD.