In the vault section, remove the text: "Vault level: x (Display user's vault level)", and add an area on top of the animated hexagonal background shape that neatly and centered displays the user's ELINT reserves in two lines, for example: "10 000 (prominent) newline ELINT (less prominent).
When a new user have joined the game, I want them to have in their inventory one level 1 pick and four level one Cypher Locks. And the first time a user places a lock in a vault slot the user must be rewarded with 10 000 ELINT (this reward is temporary and allows me to test and make a purchase in the shop).