I want to set this image to the shop's about page's background: public/spyshop/bg_about_page.jpg.
To be clear, the shop's about page won't need the hexagonal background, and the shop's logo background image can be removed.

The shop page will have the new image as a background, and on top of that a black 70 % transparent background to make the main content easily legible.

I want the new background to not repeat and fill its container vertically. On a mobile, the background is, relatively speaking, much wider than the viewport so I want the background to scroll horizontally as the user scrolls down the page. So initially, when a user lands on and is at the top of the about page, the wallpaper is aligned to the left, and as the user scrolls down, the background scrolls relative to the scrolling so that when the bottom of the about page is reached the background is aligned to the right.

On a laptop this effect will be lost seeing as no scrolling is necessary, but the game is intended for mobile so this effect will be really cool, please.