Our image is still not displaying, but let's remove the ability to click on the image to view it full size. In the details view, the image is already displayed full width (on mobile) so opening the image in a new window doesn't give you any advantage seeing as we don't have scroll to zoom etc. So to clarify, we remove the new window that displays the image completely, and clicking the image in the details window will have no effect.
While we are busy with the details screen, the level selector bar is horizontally scrollable, but on a narrow screen the far left L1 button is cut off and scrolling doesn't scroll all the way left. We had the same problem with the categories bar that we set to align left on mobiles and centered on larger screens.

And the back button sometimes floats on top of the level selector bar.  To create some extra space and avoid positioning setting of the back button, I like the idea of the back button being the first, far left button in the level selector bar, and when the level bar is scrolled horizontally, the back button remains static and always in view, but the level buttons scroll in the remaining area of the bar. And the back button can be themed to have a different appearance as the current level buttons. Is this possible?

And can the level bar be static and stay in view as the user scrolls down to the details? Being able to look at the details and then switch levels is so much easier to having to scroll back up to change the level and back down to see the new details. Would this be possible too?