did you make the changes? it didn't seem to take effect as the cards don't have backdrop blur or transparent background. and with the expanding card, that doesn't happen as the old method of hiding the carousel and then displaying the items is still active, and the up/back option becomes unresponsive when a item card has been selected so i cant get back to the main carousel view (but this doesn't need to be solved as we're going with the new expanding container approach.
If the expanding container is very complicated, which it sounds like it is, then an alternative is to, on top of the original carousel, have a borderless transparent window open that contains the new cards. We can animate zoom this new window to appear as if it is coming out of the center of the selected card (which is centered anyway after selection) and with no background it will appear as if the new cards are floating on top of the carousel. This avoids complex repositioning and resizing (I thought one could just add more content to the carousel and the selected card would autoadjust in width as a container for some of the items. Closing or going back just zooms the transparent window back into the center of the selected card.