Sorry for asking again, but did your changes take effect? For one, the background color didn't change, and the items window still has its border and background.
You said: "Implementing true infinite looping for a draggable carousel...". Is there a difference between a draggable and a scrollable carousel? The carousel doesn't have to be draggable but just scrollable if that makes a difference.

On mobile, sliding the carousel works well, and sliding on the arrows area scrolls the tod well. On laptop, with two finger type scrolling on a laptop's mouse pad, the carousel must scroll left and right if the mouse hovers above it, which is the same way the tod is scrolled.

Can you see if your changes have been made and if scrolling the carousel is doable?