there's still no background or backdrop blur. If you say the element behind the card has need to have content, does the scrolling background tod wallpaper qualify, or is that why none of my elements seem to have backdrop blur? the vault section, for example, im under the impression the hex 70% black background element has a backdrop blur, which I'm not seeing, and the same for the comms area in the control center section. when I click on a node in the scanner section's node are, the little window that opens does have backdrop blur, because there's something behind it?
Sometimes I think also the transparency can interfere and cause the background not to be displayed but fully opague would display. How can we troubleshoot our backgrounds, transparrencies, and backdrop blur effects to pinpoint which one is breaking the display?

Also, now when selecting a stack card, the individual cards window is not displaying anything except title area and the bottom buttons.