Yes, the arrow button work perfect and the animation looks awesome when switching the centered card. But the card's background color is still team themed. Let's ignore and disable any backdrop blurs in the locker section and focus on getting the color right which is more important.
And on a laptop, the vault section and the other sections have a max-width. can you apply the same max width and padding around the equipment locker section as the vault section's?

The card's resizing animation looks really cool, but it is triggered when scrolling the tod which makes it jumpy even though the carousel itself isn't being scrolled.

Then can we also see why the carousel isn't scrolling with mouse hover trackpad horizontal scrolling? We can remove the click-to-drag functionality if it can scroll with the trackpad, and there is always the left/right button to navigate the locker.