Our border colors on the cards are still unthemed. I used to work when we had css handle the and when we went the inline option to troubleshoot the backgrounds the borders stopped working. What could cause this you think?
Our card backgrounds are still the greyish/purplish color instead of the item level colors.

And our selected card still animates. But, I notice that even if not scrolling, the card wants to animate when the mouse makes the switch between hovering over the carousel or not, so if i swirl the mouse over the boundary i can get the card to animate. So it's not the scrolling perse causing the animation maybe, but that on-hover switch?